# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/goals": {
      "get": {
        "operationId": "GoalsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Own active personal goals",
        "tags": [
          "amal"
        ]
      },
      "post": {
        "operationId": "GoalsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a personal goal (max 14 active)",
        "tags": [
          "amal"
        ]
      },
      "delete": {
        "operationId": "GoalsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete one of my goals",
        "tags": [
          "amal"
        ]
      }
    },
...
    "/api/masala": {
      "post": {
        "operationId": "MasalaController_masala",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasalaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Ask a fiqh question (guests allowed)",
        "tags": [
          "masala"
        ]
      }
    },
    "/api/feedback": {
      "post": {
        "operationId": "FeedbackController_feedback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Send app feedback (guests allowed)",
        "tags": [
          "feedback"
        ]
      }
    },
    "/api/enroll": {
      "post": {
        "operationId": "EnrollController_enroll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Enroll in a course (idempotent)",
        "tags": [
          "enroll"
        ]
      },
      "patch": {
        "operationId": "EnrollController_progress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Save course progress JSON",
        "tags": [
          "enroll"
        ]
      }
    },
    "/api/quiz-attempt": {
      "post": {
        "operationId": "QuizAttemptController_attempt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizAttemptDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Record a quiz attempt (login)",
        "tags": [
          "quiz-attempt"
        ]
      }
    },
    "/api/reminders": {
      "get": {
        "operationId": "RemindersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Own reminders",
        "tags": [
          "reminders"
        ]
      },
      "patch": {
        "operationId": "RemindersController_read",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadReminderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mark a reminder read",
        "tags": [
          "reminders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sunnah Life API",
    "description": "দাওয়াতুস সুন্নাহ তারবিয়াত প্ল্যাটফর্ম — auth (OTP + JWT), muhasaba diary, usrah & dawah engine, weekly reviews, assessments, admin console. Every data route is enforced by PostgreSQL Row-Level Security (gender / usrah / downline scoping); auth via `Authorization: Bearer <accessToken>` or the HttpOnly sl_access cookie.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sl_access"
      }
    },
    "schemas": {
...
      "GoalCreateDto": {
        "type": "object",
        "properties": {
          "amalKey": {
            "type": "string",
            "example": "tahajjud"
          },
          "title": {
            "type": "string",
            "example": "তাহাজ্জুদ নিয়মিত করা"
          },
          "note": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "example": "2025-06-01"
          }
        },
        "required": [
          "amalKey",
          "title",
          "startDate"
        ]
      },
...
      "MasalaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "তানভীর হোসেন"
          },
          "phone": {
            "type": "string",
            "example": "01000000007"
          },
          "question": {
            "type": "string",
            "example": "মাসনূন আযকার কি সব একসাথে পড়া যায়?"
          }
        },
        "required": [
          "name",
          "question"
        ]
      },
      "FeedbackDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "EnrollDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "example": "course-tawheed-101"
          }
        },
        "required": [
          "courseId"
        ]
      },
      "EnrollProgressDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "example": "course-tawheed-101"
          },
          "progressJson": {
            "type": "string",
            "example": "{\"lessonIndex\":2,\"completed\":[1,2]}"
          }
        },
        "required": [
          "courseId",
          "progressJson"
        ]
      },
      "QuizAttemptDto": {
        "type": "object",
        "properties": {
          "quizId": {
            "type": "string",
            "example": "quiz-fiqh-basics"
          },
          "score": {
            "type": "number",
            "minimum": 0
          },
          "total": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "quizId",
          "score",
          "total"
        ]
      },
      "ReadReminderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}